review fixes.
